Move mail notification logic to service layer
